pyproject.toml
[tool.poetry]
name = "aid-web-backend-v2"
version = "0.1.0"
description = ""
authors = ["sihyeong671 <bshlab671@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
django-seed = "^0.3.1"
drf-spectacular = "^0.27.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
